feat: 新增 minio 单机

See merge request szyl/zy/shuyan/datasophon!6

+++ datasophon-common/src/main/java/com/datasophon/common/model/ServiceConfig.java
@@ -41,9 +41,9 @@ public class ServiceConfig implements Serializable {
     
     private Object defaultValue;
     
-    private Integer minValue;
+    private Long minValue;
     
-    private Integer maxValue;
+    private Long maxValue;
     
     private String unit;
